Split up tracing.py into separate modules
