+++ tests/integrations/opentelemetry/test_span_processor.py
@@ -9,7 +9,7 @@ from sentry_sdk.integrations.opentelemetry.span_processor import (
 from sentry_sdk.tracing import Span, Transaction
 
 from opentelemetry.trace import SpanKind, SpanContext
-from sentry_sdk.tracing_utils import extract_sentrytrace_data
+from sentry_sdk.tracing.utils import extract_sentrytrace_data
 
 
 def test_is_sentry_span():
